requirements.txt
discord.py>=2.4.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
tzdata>=2024.1
audioop-lts>=0.2.1
